refactor: 불필요한 메서드 삭제 및 단위테스트 추가
