+++ src/test/java/domain/participant/info/NameTest.java
@@ -1,12 +1,10 @@
-package domain.participant;
+package domain.participant.info;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
-
-import domain.participant.info.Name;
 
 public class NameTest {
